split synthesis engine mora test
